requirements.txt
selenium>=4.9.0
webdriver-manager>=3.8.6
python-dotenv>=0.19.0
requests>=2.26.0
colorlog>=6.7.0
openai>=1.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
